+++ src/main/java/gyro/azure/keyvault/KeyVaultSecretResource.java
@@ -1,5 +1,7 @@
 package gyro.azure.keyvault;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
 
 import com.microsoft.azure.keyvault.models.KeyVaultErrorException;
@@ -8,6 +10,7 @@ import com.microsoft.azure.keyvault.requests.SetSecretRequest;
 import com.microsoft.azure.keyvault.requests.UpdateSecretRequest;
 import com.microsoft.azure.management.keyvault.Vault;
 import com.psddev.dari.util.ObjectUtils;
+import gyro.azure.AzureResource;
 import gyro.azure.Copyable;
 import gyro.core.GyroUI;
 import gyro.core.Type;
@@ -43,15 +46,17 @@ import gyro.core.validation.Required;
  *     end
  */
 @Type("key-vault-secret")
-public class KeyVaultSecretResource extends KeyVaultInnerResource implements Copyable<SecretBundle> {
+public class KeyVaultSecretResource extends AzureResource implements Copyable<SecretBundle> {
 
     private String name;
+    private KeyVaultResource vault;
     private String value;
     private KeyVaultSecretAttribute attribute;
     private String contentType;
     private String id;
     private String kid;
     private String identifier;
+    private Map<String, String> tags;
 
     /**
      * The name of the secret. (Required)
@@ -63,6 +68,18 @@ public class KeyVaultSecretResource extends KeyVaultInnerResource implements Cop
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    /**
+     * The key vault under which the secret is going to be created. (Required)
+     */
+    @Required
+    public KeyVaultResource getVault() {
+        return vault;
+    }
+
+    public void setVault(KeyVaultResource vault) {
+        this.vault = vault;
     }
 
     /**
@@ -141,6 +158,22 @@ public class KeyVaultSecretResource extends KeyVaultInnerResource implements Cop
         this.identifier = identifier;
     }
 
+    /**
+     * Tags for the secret.
+     */
+    @Updatable
+    public Map<String, String> getTags() {
+        if (tags == null) {
+            tags = new HashMap<>();
+        }
+
+        return tags;
+    }
+
+    public void setTags(Map<String, String> tags) {
+        this.tags = tags;
+    }
+
     @Override
     public void copyFrom(SecretBundle secret) {
         setName(secret.secretIdentifier().name());
@@ -161,7 +194,7 @@ public class KeyVaultSecretResource extends KeyVaultInnerResource implements Cop
 
     @Override
     public boolean refresh() {
-        Vault vault = getKeyVault();
+        Vault vault = getVault().getKeyVault();
         SecretBundle secret;
 
         try {
@@ -186,7 +219,7 @@ public class KeyVaultSecretResource extends KeyVaultInnerResource implements Cop
 
     @Override
     public void create(GyroUI ui, State state) throws Exception {
-        Vault vault = getKeyVault();
+        Vault vault = getVault().getKeyVault();
 
         SetSecretRequest.Builder builder = new SetSecretRequest.Builder(vault.vaultUri(), getName(), getValue());
 
@@ -208,7 +241,7 @@ public class KeyVaultSecretResource extends KeyVaultInnerResource implements Cop
     @Override
     public void update(
         GyroUI ui, State state, Resource current, Set<String> changedFieldNames) throws Exception {
-        Vault vault = getKeyVault();
+        Vault vault = getVault().getKeyVault();
 
         UpdateSecretRequest.Builder builder = new UpdateSecretRequest.Builder(getId());
 
@@ -221,7 +254,7 @@ public class KeyVaultSecretResource extends KeyVaultInnerResource implements Cop
 
     @Override
     public void delete(GyroUI ui, State state) throws Exception {
-        Vault vault = getKeyVault();
+        Vault vault = getVault().getKeyVault();
 
         vault.client().deleteSecret(vault.vaultUri(), getName());
     }
